bugfixes for tab4, refactoring base pages for tab1 and tab4
