+++ webinterface/pages/base_pages/tab1_results.py
@@ -4,19 +4,12 @@ from typing import Callable
 
 import streamlit as st
 import pandas as pd
-import re
 
-from proteobench.plotting.plot_quant import PlotDataPoint
-from st_aggrid import GridOptionsBuilder, AgGrid, JsCode
+from .resulttable import configure_aggrid, render_aggrid, prepare_display_dataframe
+from .metricplot import render_metric_plot
+
 
 from .filter import filter_data_using_slider
-
-# === Table Color Constants ===
-COLOR_IDENTIFIER = "#F0F2F6"
-COLOR_PARAMETER = "#FFFFFF"
-COLOR_RESULT = "#F0F2F6"
-COLOR_TECHNICAL = "#FFFFFF"
-COLOR_ADDITIONAL = "#F0F2F6"
 
 
 def initialize_main_slider(slider_id_uuid: str, default_val_slider: float) -> None:
@@ -149,13 +142,31 @@ def display_existing_results(variables_quant, ionmodule) -> None:
     initialize_and_filter_data(variables_quant, ionmodule)
     data_points_filtered = variables_quant.filtered_data
 
-    metric = display_metric_selector()
-    highlight_point_id = render_metric_plot(data_points_filtered, variables_quant, metric)
+    metric = display_metric_selector(variables_quant)
+
+    # prepare plot key explicitly for tab 1
+    key = variables_quant.result_plot_uuid
+    if key not in st.session_state.keys():
+        st.session_state[key] = uuid.uuid4()
+    _id_of_key = st.session_state[key]
+
+    highlight_point_id = render_metric_plot(
+        data_points_filtered,
+        metric,
+        label=st.session_state[st.session_state[variables_quant.selectbox_id_uuid]],
+        key=_id_of_key,
+    )
 
     df_display = prepare_display_dataframe(data_points_filtered, highlight_point_id)
     grid_options = configure_aggrid(df_display)
 
-    render_aggrid(df_display, grid_options)
+    # prepare df key explicitly for tab 1
+    key = variables_quant.table_id_uuid
+    if key not in st.session_state.keys():
+        st.session_state[key] = uuid.uuid4()
+    _id_of_key = st.session_state[key]
+
+    render_aggrid(df_display, grid_options, key)
     offer_download(df_display)
     display_download_section(variables_quant=variables_quant)
 
@@ -175,223 +186,17 @@ def initialize_and_filter_data(variables_quant, ionmodule):
     )
 
 
-def display_metric_selector() -> str:
+def display_metric_selector(variables_quant) -> str:
+    key = variables_quant.metric_selector_uuid
+    if key not in st.session_state.keys():
+        st.session_state[key] = uuid.uuid4()
+    _id_of_key = st.session_state[key]
+
     return st.radio(
         "Select metric to plot",
         options=["Median", "Mean"],
         help="Toggle between median and mean absolute difference metrics.",
-    )
-
-
-def render_metric_plot(data: pd.DataFrame, variables_quant, metric: str) -> str | None:
-    """
-    Displays the metric plot and returns the ProteoBench ID of the selected point (if any).
-
-    Parameters
-    ----------
-    data : pd.DataFrame
-        The filtered dataset to plot.
-
-    variables_quant : object
-        Contains session state and selectbox identifier.
-
-    metric : str
-        Metric to plot ("Median" or "Mean").
-
-    Returns
-    -------
-    str or None
-        ProteoBench ID of the selected data point, if any.
-    """
-
-    if len(data) == 0:
-        st.error("No datapoints available for plotting", icon="🚨")
-        return None
-
-    highlight_point_id = None
-    try:
-        fig_metric = PlotDataPoint.plot_metric(
-            data,
-            label=st.session_state[st.session_state[variables_quant.selectbox_id_uuid]],
-            metric=metric,
-        )
-        event_dict = st.plotly_chart(
-            fig_metric,
-            use_container_width=True,
-            on_select="rerun",
-            selection_mode="points",
-        )
-        selected_point = (
-            event_dict["selection"]["points"][0]
-            if "selection" in event_dict and "points" in event_dict["selection"] and event_dict["selection"]["points"]
-            else None
-        )
-        if selected_point:
-            hover = selected_point.get("hovertext", "")
-            match = re.search(r"ProteoBench ID: ([^\s<]+)", hover)
-            if match:
-                highlight_point_id = match.group(1)
-
-    except Exception as e:
-        st.error(f"Unable to plot the datapoints: {e}", icon="🚨")
-
-    return highlight_point_id
-
-
-def prepare_display_dataframe(df: pd.DataFrame, highlight_id: str | None) -> pd.DataFrame:
-    """
-    Prepares the DataFrame for display, including column filtering, ordering,
-    row highlighting, and numeric formatting.
-
-    Parameters
-    ----------
-    df : pd.DataFrame
-        The filtered dataset for display.
-
-    highlight_id : str or None
-        The ProteoBench ID to highlight (adds a marker in the 'selected' column).
-
-    Returns
-    -------
-    pd.DataFrame
-        A formatted and sorted DataFrame ready for rendering.
-    """
-    df = df.copy()
-    df["selected"] = df["id"].apply(lambda x: "➡️" if x == highlight_id else "")
-
-    identifier_cols = ["selected", "id"]
-    parameter_cols = [
-        "software_name",
-        "software_version",
-        "search_engine",
-        "search_engine_version",
-        "ident_fdr_psm",
-        "ident_fdr_protein",
-        "ident_fdr_peptide",
-        "enable_match_between_runs",
-        "precursor_mass_tolerance",
-        "fragment_mass_tolerance",
-        "enzyme",
-        "allowed_miscleavages",
-        "min_peptide_length",
-        "max_peptide_length",
-        "fixed_mods",
-        "variable_mods",
-        "max_mods",
-        "min_precursor_charge",
-        "max_precursor_charge",
-        "quantification_method",
-        "protein_inference",
-        "abundance_normalization_ions",
-        "submission_comments",
-    ]
-    result_cols = ["median_abs_epsilon", "mean_abs_epsilon", "nr_prec", "results"]
-    technical_cols = [
-        "proteobench_version",
-        "intermediate_hash",
-        "hover_text",
-        "color",
-        "old_new",
-        "is_temporary",
-        "comments",
-        "scatter_size",
-    ]
-
-    # Define display column order
-    cols = identifier_cols + parameter_cols + result_cols + technical_cols
-    cols = [col for col in cols if col in df.columns]
-    additional_cols = [col for col in df.columns if col not in cols]
-    # remove boring columns
-    cols = [
-        col for col in cols if col not in ["comments", "scatter_size", "old_new", "is_temporary", "color", "hover_text"]
-    ]
-    df = df[cols + additional_cols]
-
-    # Clean up values
-    df["results"] = df["results"].apply(str)
-    numeric_cols = df.select_dtypes(include=["float64", "int64"]).columns
-    df[numeric_cols] = df[numeric_cols].round(3)
-    df.sort_values(by="id", inplace=True)
-
-    return df
-
-
-def configure_aggrid(df: pd.DataFrame):
-    """
-    Configures the styling and options for AgGrid based on column category.
-
-    Parameters
-    ----------
-    df : pd.DataFrame
-        The display-ready DataFrame.
-
-    Returns
-    -------
-    dict
-        AgGrid gridOptions dictionary.
-    """
-    gb = GridOptionsBuilder.from_dataframe(df)
-    identifier_cols = ["selected", "id"]
-    parameter_cols = [
-        "software_name",
-        "software_version",
-        "search_engine",
-        "search_engine_version",
-        "ident_fdr_psm",
-        "ident_fdr_protein",
-        "ident_fdr_peptide",
-        "enable_match_between_runs",
-        "precursor_mass_tolerance",
-        "fragment_mass_tolerance",
-        "enzyme",
-        "allowed_miscleavages",
-        "min_peptide_length",
-        "max_peptide_length",
-        "fixed_mods",
-        "variable_mods",
-        "max_mods",
-        "min_precursor_charge",
-        "max_precursor_charge",
-        "quantification_method",
-        "protein_inference",
-        "abundance_normalization_ions",
-        "submission_comments",
-    ]
-    result_cols = ["median_abs_epsilon", "mean_abs_epsilon", "nr_prec", "results"]
-    technical_cols = [
-        "proteobench_version",
-        "intermediate_hash",
-        "hover_text",
-        "color",
-        "old_new",
-        "is_temporary",
-        "comments",
-        "scatter_size",
-    ]
-
-    for col in df.columns:
-        if col in identifier_cols:
-            gb.configure_column(col, cellStyle=get_style_js(COLOR_IDENTIFIER))
-        elif col in parameter_cols:
-            gb.configure_column(col, cellStyle=get_style_js(COLOR_PARAMETER))
-        elif col in result_cols:
-            gb.configure_column(col, cellStyle=get_style_js(COLOR_RESULT))
-        elif col in technical_cols:
-            gb.configure_column(col, cellStyle=get_style_js(COLOR_TECHNICAL))
-        else:
-            gb.configure_column(col, cellStyle=get_style_js(COLOR_ADDITIONAL))
-
-    return gb.build()
-
-
-def render_aggrid(df: pd.DataFrame, grid_options):
-    AgGrid(
-        df,
-        gridOptions=grid_options,
-        theme="alpine",
-        fit_columns_on_grid_load=False,
-        height=600,
-        allow_unsafe_jscode=True,
+        key=_id_of_key,
     )
 
 
@@ -409,33 +214,6 @@ def offer_download(df: pd.DataFrame, filename: str = "quantification_results.csv
     """
     csv_data = df.to_csv(index=False).encode("utf-8")
     st.download_button(label="📥 Download table as CSV", data=csv_data, file_name=filename, mime="text/csv")
-
-
-def get_style_js(bg_color: str) -> JsCode:
-    """
-    Generates JavaScript for styling cells with a background color.
-
-    Parameters
-    ----------
-    bg_color : str
-        Hex color string to use as the background.
-
-    Returns
-    -------
-    JsCode
-        A JavaScript code block that defines the style.
-    """
-    return JsCode(
-        f"""
-    function(params) {{
-        return {{
-            'backgroundColor': '{bg_color}',
-            'color': 'black',
-            'fontWeight': 'normal'
-        }}
-    }}
-    """
-    )
 
 
 def initialize_main_data_points(all_datapoints: str, obtain_all_data_points: Callable) -> None:
