+++ webinterface/pages/pages_variables/Quant/lfq_DDA_ion_QExactive_variables.py
@@ -45,6 +45,11 @@ class VariablesDDAQuant:
     slider_id_uuid: str = "slider_id"
     download_selector_id_uuid: str = "download_selector_id"
     table_id_uuid: str = "table_id"
+    table_new_results_uuid: str = "table_new_results_uuid"
+    result_plot_uuid: str = "result_figure_uuid"
+    result_submitted_plot_uuid: str = "result_submitted_figure_uuid"
+    metric_selector_uuid: str = "metric_selector_uuid"
+    metric_selector_submitted_uuid: str = "metric_selector_submitted_uuid"
 
     metric_plot_labels: List[str] = field(
         default_factory=lambda: [
