+++ webinterface/pages/pages_variables/Quant/lfq_DIA_ion_Astral_variables.py
@@ -49,6 +49,11 @@ class VariablesDIAQuantAstral:
     slider_id_uuid: str = "slider_id_dia_quant_Astral"
     download_selector_id_uuid: str = "download_selector_id_dia_quant_Astral"
     table_id_uuid: str = "table_id_dia_quant_Astral"
+    table_new_results_uuid: str = "table_new_results_uuid_dia_quant_Astral"
+    result_plot_uuid: str = "result_figure_uuid_dia_quant_Astral"
+    result_submitted_plot_uuid: str = "result_submitted_figure_uuid_dia_quant_Astral"
+    metric_selector_uuid: str = "metric_selector_uuid_dia_quant_Astral"
+    metric_selector_submitted_uuid: str = "metric_selector_submitted_uuid_dia_quant_Astral"
 
     metric_plot_labels: List[str] = field(
         default_factory=lambda: [
