+++ webinterface/pages/pages_variables/Quant/lfq_DIA_ion_singlecell_variables.py
@@ -48,6 +48,11 @@ class VariablesDIAQuantSC:
     slider_id_uuid: str = "slider_id_dia_quant_singlecell"
     download_selector_id_uuid: str = "download_selector_id_dia_quant_singlecell"
     table_id_uuid: str = "table_id_dia_quant_singlecell"
+    table_new_results_uuid: str = "table_new_results_uuid_dia_quant_singlecell"
+    result_plot_uuid: str = "result_figure_uuid_dia_quant_singlecell"
+    result_submitted_plot_uuid: str = "result_submitted_figure_uuid_dia_quant_singlecell"
+    metric_selector_uuid: str = "metric_selector_uuid_dia_quant_singlecell"
+    metric_selector_submitted_uuid: str = "metric_selector_submitted_uuid_dia_quant_singlecell"
 
     metric_plot_labels: List[str] = field(
         default_factory=lambda: [
